moved identifier checks into Tokenizer
